requirements.txt
backcall==0.2.0
debugpy==1.5.1
decorator==5.1.0
entrypoints==0.3
jedi==0.18.0
kafka-python==2.0.2
matplotlib-inline==0.1.3
nest-asyncio==1.5.1
numpy==1.21.4
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.22
ptyprocess==0.7.0
Pygments==2.10.0
python-dateutil==2.8.2
pyzmq==22.3.0
scipy==1.7.2
six==1.16.0
termcolor==1.1.0
tornado==6.1
tqdm==4.62.3
traitlets==5.1.1
wcwidth==0.2.5
sklearn
